Move expression widget into subfolder
