+++ frontend/src/metabase/query_builder/components/expressions/CombineColumns/CombineColumns.tsx
@@ -6,7 +6,7 @@ import { isNotNull } from "metabase/lib/types";
 import { Box, Button, Flex, Icon, Stack } from "metabase/ui";
 import * as Lib from "metabase-lib";
 
-import { ExpressionWidgetHeader } from "../ExpressionWidgetHeader";
+import { ExpressionWidgetHeader } from "../ExpressionWidget";
 
 import { ColumnAndSeparatorRow } from "./ColumnAndSeparatorRow";
 import { Example } from "./Example";
