+++ frontend/src/metabase/query_builder/components/expressions/ExpressionWidget/index.ts
@@ -1,0 +1,2 @@
+export * from "./ExpressionWidget";
+export * from "./ExpressionWidgetHeader";
