+++ frontend/src/metabase/query_builder/components/expressions/ExtractColumn/ExtractColumn.tsx
@@ -6,7 +6,7 @@ import { getExample } from "metabase/querying/drills/utils/column-extract-drill"
 import { Box, Button, Flex, Stack, Text, Title } from "metabase/ui";
 import * as Lib from "metabase-lib";
 
-import { ExpressionWidgetHeader } from "../ExpressionWidgetHeader";
+import { ExpressionWidgetHeader } from "../ExpressionWidget";
 
 import styles from "./ExtractColumn.module.css";
 import { getName } from "./util";
